Shrink output apk size.
This does reduce it by about 30% . (27M -> 17M)

It could be more effective if there was a way to enable stronger zip compression.
But I haven't been able to find a way to do so yet.

For now maybe distribute the apk gzipped (apk.gz) to improve compression.
This reduces size by about another 50 % (output ~ 8.7M)